backend/requirements.txt
Flask==2.0.1
torch==1.9.0
torchvision==0.10.0
torchinfo==1.5.2
Pillow==8.3.1
gunicorn==20.1.0